shelby_as_a_service/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.11.0
blis==0.7.11
bs4==0.0.1
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
confection==0.1.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.2
debugpy==1.8.0
decorator==5.1.1
discord==2.3.2
discord.py==2.3.2
distro==1.8.0
dnspython==2.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
gradio==4.2.0
gradio_client==0.7.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.1
idna==3.4
importlib-resources==6.1.1
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
langchain==0.0.335
langcodes==3.3.0
langsmith==0.0.63
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.0.1
multidict==6.0.4
murmurhash==1.0.10
mypy==1.7.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.2
openai==1.2.3
openapi-schema-pydantic==1.2.4
orjson==3.9.10
packaging==23.2
pandas==2.1.3
parso==0.8.3
pathspec==0.11.2
pathy==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==4.0.0
preshed==3.0.9
prompt-toolkit==3.0.40
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.0
pydantic_core==2.14.1
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.23.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
style==1.1.6
tenacity==8.2.3
thinc==8.1.12
tiktoken==0.5.1
tomlkit==0.12.0
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
wasabi==1.1.2
wcwidth==0.2.9
websockets==11.0.3
wget==3.2
yarl==1.9.2
